pyproject.toml
[tool.poetry]
name = "imap bankoff bot"
version = "0.1.0"
description = ""
authors = ["Malvere <gatling1404@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.19"
pycryptodome = "^3.14.1"

[tool.poetry.dev-dependencies]
ipykernel = "^6.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
